+++ src/config.ts
@@ -21,10 +21,13 @@ import { createAPIClient } from './client';
  * `instance.config` in a UI.
  */
 export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
-  clientId: {
+  userId: {
     type: 'string',
   },
-  clientSecret: {
+  hostname: {
+    type: 'string',
+  },
+  apiKey: {
     type: 'string',
     mask: true,
   },
@@ -38,12 +41,17 @@ export interface IntegrationConfig extends IntegrationInstanceConfig {
   /**
    * The provider API client ID used to authenticate requests.
    */
-  clientId: string;
+  userId: string;
 
   /**
    * The provider API client secret used to authenticate requests.
    */
-  clientSecret: string;
+  apiKey: string;
+
+  /**
+   * The provider hostname.
+   */
+  hostname: string;
 }
 
 export async function validateInvocation(
@@ -51,9 +59,9 @@ export async function validateInvocation(
 ) {
   const { config } = context.instance;
 
-  if (!config.clientId || !config.clientSecret) {
+  if (!config.userId || !config.apiKey || !config.hostname) {
     throw new IntegrationValidationError(
-      'Config requires all of {clientId, clientSecret}',
+      'Config requires all of {userId, apiKey, hostname}',
     );
   }